move auth0 file to functions
